table printer: expand table
When we are given a with that is larger than the content we have,
expand the space between columns, within reason. We only make the
table at most 1.5 times as wide as it would be unexpanded.

And make flatpak_table_printer_print_full return not just the table
height, but also the width.

Update all callers.