Add custom, high-perfomance prune implementation
NOTE: This is currently WIP and doesn't do proper locking, so
the actual deletion of objects is disabled. Its mainly useful for testing.

This is basically an open-coded version of ostree_repo_prune() that is
faster so that we can prune larger repos (like flathub) in a realistic
timeframe.

The primary reason it is faster is that it creates and uses a .commitmeta2
file for each commit with information about what objects are reachable
from that commit. This means incremental prunes need only traverse over
new commits.

Secondly, it uses the variant parser compiled accessors for the various
GVariants that are involved in the prune which is quite a bit faster,
especially if the repo is very large.